chore(actions): simplifies typings and middlewares
