+++ server-fetch/public/index.ts
@@ -3,24 +3,16 @@ import { getEORIValidation } from "#models/eori-validation";
 import { getDirigeantsRNE } from "#models/rne/dirigeants";
 import { getRNEObservations } from "#models/rne/observations";
 import { buildAndVerifyTVA } from "#models/tva/verify";
-import { withErrorHandler } from "#utils/server-side-helper/with-error-handler";
+import { withErrorHandler } from "../middlewares";
 
-export const getRNEObservationsFetcher = withErrorHandler((siren) =>
-  getRNEObservations(siren)
+export const getRNEObservationsFetcher = withErrorHandler(getRNEObservations);
+
+export const getDirigeantsRNEFetcher = withErrorHandler(getDirigeantsRNE);
+
+export const getAssociationFromSlugFetcher = withErrorHandler(
+  getAssociationFromSlug
 );
 
-export const getDirigeantsRNEFetcher = withErrorHandler((siren) =>
-  getDirigeantsRNE(siren)
-);
+export const getEORIValidationFetcher = withErrorHandler(getEORIValidation);
 
-export const getAssociationFromSlugFetcher = withErrorHandler((siren) =>
-  getAssociationFromSlug(siren)
-);
-
-export const getEORIValidationFetcher = withErrorHandler((siret) =>
-  getEORIValidation(siret)
-);
-
-export const buildAndVerifyTVAFetcher = withErrorHandler((siren) =>
-  buildAndVerifyTVA(siren)
-);
+export const buildAndVerifyTVAFetcher = withErrorHandler(buildAndVerifyTVA);
